move to substrate watcher moudle #136
